# api/openapi.yaml
openapi: 3.0.0
info:
  title: HR Recruiter API
  version: 1.0.0
  description: |
    Полная OpenAPI спецификация для платформы автоматизации.
    Включает все основные сущности, бизнес-операции, заглушки для AI и медиа, а также примеры для быстрой интеграции.
servers:
  - url: http://localhost:8080/api/v1
    description: Local server
  - url: https://api.hr-recruiter.com/api/v1
    description: Production server
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags:
        - Team & Users
      summary: Список пользователей
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Номер страницы (начиная с 0)
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Размер страницы
        - in: query
          name: sort
          schema:
            type: string
            default: createdAt
          description: Поле для сортировки
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersPaginatedResponse'
    post:
      operationId: createUser
      tags:
        - Team & Users
      summary: Создать пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      operationId: getUser
      tags:
        - Team & Users
      summary: Получить пользователя
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      operationId: updateUser
      tags:
        - Team & Users
      summary: Обновить пользователя
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Пользователь обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      operationId: deleteUser
      tags:
        - Team & Users
      summary: Удалить пользователя по ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Пользователь успешно удален
...
components:
  schemas:
...
    BaseEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, createdAt]
    
    # Общий класс для пагинированных ответов
    PaginatedResponse:
      type: object
      description: "Базовый класс для всех пагинированных ответов"
      properties:
        content:
          type: array
          description: "Список элементов на текущей странице"
          items:
            type: object
        totalElements:
          type: integer
          format: int64
          description: "Общее количество элементов"
          example: 150
        totalPages:
          type: integer
          format: int64
          description: "Общее количество страниц"
          example: 15
        number:
          type: integer
          format: int64
          description: "Номер текущей страницы (начиная с 0)"
          example: 0
        size:
          type: integer
          format: int64
          description: "Размер страницы"
          example: 10
      required: [content, totalElements, totalPages, number, size]
    
    # Пагинированные ответы для разных сущностей
...
    UsersPaginatedResponse:
      $ref: '#/components/schemas/PaginatedResponse'
      description: "Пагинированный список пользователей"
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/User'
...
    BaseUserFields:
      type: object
      properties:
        firstName:
          type: string
          description: "Имя пользователя"
        lastName:
          type: string
          description: "Фамилия пользователя"
        email:
          type: string
        role:
          $ref: '#/components/schemas/RoleEnum'
        avatarUrl:
          type: string
        language:
          type: string
        phone:
          type: string
    
    # Основные схемы
...
    User:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - $ref: '#/components/schemas/BaseUserFields'
      required: [id, firstName, lastName, email, role, createdAt]
    UserCreateRequest:
      type: object
      properties:
        firstName:
          type: string
          description: "Имя пользователя"
        lastName:
          type: string
          description: "Фамилия пользователя"
        email:
          type: string
        role:
          $ref: '#/components/schemas/RoleEnum'
        avatarUrl:
          type: string
        language:
          type: string
        phone:
          type: string
        password:
          type: string
      required: [firstName, lastName, email, role, password]
    UserUpdateRequest:
      $ref: '#/components/schemas/BaseUserFields'
...
    RoleEnum:
      type: string
      enum:
        - admin
        - recruiter
        - viewer
    
    # Stats
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
security:
  - basicAuth: []
...